video_platform/requirements.txt
django==6.0.4
djangorestframework==3.17.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
gunicorn==25.3.0
whitenoise==6.12.0
django-extensions==4.1
werkzeug==3.1.8
whitenoise==6.12.0